dags/extract_test_cli/requirements.txt
python-dotenv==0.21.1
fastavro
flickrapi
praw